requirements.txt
agno
google-generativeai
python-dotenv
pandas
matplotlib
google-genai